+++ pixi.toml
@@ -16,7 +16,9 @@ matplotlib = ">=3.9.1,<3.10"
 more-itertools = ">=10.3.0,<10.4"
 cvxpy = ">=1.5.2,<1.6"
 pandas = ">=2.2.2,<2.3"
+seaborn = ">=0.13.2,<0.14"
 
 [pypi-dependencies]
 pyteomics = "*"
 xlsxwriter = "*"
+adjusttext = "*"
