--- tandem_tango/constants.py
@@ -1,1 +1,0 @@
-consta
