+++ tandem_tango/deprecated/CDF_parallel.py
