+++ tandem_tango/deprecated/constants.py
@@ -1,0 +1,5 @@
+# Constants
+NS = {'mzml': 'http://psi.hupo.org/ms/mzml',
+      'ms_level': 'MS:1000511',
+      "scan_start_time" : 'MS:1000016', # in minutes
+      "isolation_window_target_mz" : 'MS:1000827'}
