+++ tandem_tango/deprecated/gammasCalc.py
